fix(auth): enforce admin permissions per-route, not on the /api mount
The admin routers were mounted as app.use('/api', authMiddleware,
requirePermission('X'), router). Because that puts the gate on the shared
'/api' prefix, every admin request runs the gate chain in mount order — so a
request to e.g. /api/admin/crawlers was first blocked by the admin.context-plugins
gate. Result: only wildcard ('*') users could reach any admin endpoint; granular
admin roles (the whole point of the configurable mapping) never worked.

Move each gate onto the individual routes inside the routers (the pattern admin.js
already uses), and drop the mount-level gates from app.js. Also split the Express
app into createApp() (app.js) so it can be imported by tests without binding a port.
